requirements.txt
django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-environ
whitenoise
gunicorn
django-cors-headers
Pillow
requests
